# src/app/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dummy Backend API",
    "description": "API documentation for the Dummy Backend system - managing HR, Payroll, Operations, Inventory, and Fleet Management.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Base URL"
    }
  ],
...
  "paths": {
    "/clean/hr_employees": {
      "get": {
        "tags": ["Employees"],
        "summary": "Get all employees or a specific employee",
        "description": "Returns a list of all employees or a specific employee by employee number",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Employee number to fetch specific employee",
            "required": false,
            "schema": { "type": "string" },
            "example": "EMP-2024-NKFN57"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    { "$ref": "#/components/schemas/Employee" },
                    { "$ref": "#/components/schemas/EmployeeList" }
                  ]
                },
                "examples": {
                  "single": {
                    "summary": "Single employee",
                    "value": {
                      "employeeNumber": "EMP-2021-OPS-007",
                      "firstName": "Mario",
                      "middleName": "Ramos",
                      "lastName": "Perez",
                      "phone": "+639389025401",
                      "position": "Driver",
                      "barangay": "Santo Niño",
                      "zipCode": "1100",
                      "departmentId": 4,
                      "department": "Operations"
                    }
                  },
                  "list": {
                    "summary": "Employee list",
                    "value": {
                      "employees": [
                        {
                          "employeeNumber": "EMP-2021-OPS-007",
                          "firstName": "Mario",
                          "middleName": "Ramos",
                          "lastName": "Perez",
                          "phone": "+639389025401",
                          "position": "Driver",
                          "barangay": "Santo Niño",
                          "zipCode": "1100",
                          "departmentId": 4,
                          "department": "Operations"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Employee not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
    "/clean/buses": {
      "get": {
        "tags": ["Buses"],
        "summary": "Get all buses or a specific bus",
        "description": "Returns bus fleet information",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Bus ID to fetch specific bus",
            "required": false,
            "schema": { "type": "string" },
            "example": "BUS001"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    { "$ref": "#/components/schemas/Bus" },
                    { "$ref": "#/components/schemas/BusList" }
                  ]
                },
                "example": {
                  "id": 5,
                  "bus_id": "BUS005",
                  "license_plate": "PLATE-0005",
                  "body_number": "BODY-0005",
                  "type": "AIRCONDITIONED",
                  "capacity": 41
                }
              }
            }
          },
          "404": { "description": "Bus not found" },
          "500": { "description": "Internal server error" }
        }
      },
      "post": {
        "tags": ["Buses"],
        "summary": "Create a new bus",
        "description": "Add a new bus to the fleet",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BusCreate" },
              "example": {
                "license_plate": "PLATE-0026",
                "body_number": "BODY-0026",
                "type": "AIRCONDITIONED",
                "capacity": 45
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Bus created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Bus" }
              }
            }
          },
          "400": { "description": "Invalid request" },
          "500": { "description": "Internal server error" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Employee": {
        "type": "object",
        "properties": {
          "employeeNumber": { "type": "string", "example": "EMP-2021-OPS-007" },
          "firstName": { "type": "string", "example": "Mario" },
          "middleName": { "type": "string", "example": "Ramos" },
          "lastName": { "type": "string", "example": "Perez" },
          "phone": { "type": "string", "example": "+639389025401" },
          "position": { "type": "string", "example": "Driver" },
          "barangay": { "type": "string", "example": "Santo Niño" },
          "zipCode": { "type": "string", "example": "1100" },
          "departmentId": { "type": "integer", "example": 4 },
          "department": { "type": "string", "example": "Operations" }
        }
      },
      "EmployeeList": {
        "type": "object",
        "properties": {
          "employees": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Employee" }
          }
        }
      },
      "Bus": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 5 },
          "bus_id": { "type": "string", "example": "BUS005" },
          "license_plate": { "type": "string", "example": "PLATE-0005" },
          "body_number": { "type": "string", "example": "BODY-0005" },
          "type": { "type": "string", "enum": ["AIRCONDITIONED", "NONAIRCON"], "example": "AIRCONDITIONED" },
          "capacity": { "type": "integer", "example": 41 }
        }
      },
      "BusList": {
        "type": "object",
        "properties": {
          "buses": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Bus" }
          }
        }
      },
      "BusCreate": {
        "type": "object",
        "required": ["license_plate", "body_number"],
        "properties": {
          "license_plate": { "type": "string" },
          "body_number": { "type": "string" },
          "type": { "type": "string", "enum": ["AIRCONDITIONED", "NONAIRCON"] },
          "capacity": { "type": "integer", "default": 45 }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      }
    }
  }
}